refactor: extract shared ATS board core into scripts/ats_boards.py
